Porosity regions implementation
Added the possibility to define a value of the porosity in regions that are not in the optimization domain. This means that now we can simulate the presence of porous regions. Obviously the range of porosity depends on alpha formulation, but being able to vary gamma, every porosity value can be reached if tuned properly.