Bugfix + New Feature: Real Stationary Solution implemented
The static solution has been corrected and correctly implemented and not just approximated using a big final time.
The static solver methods have been implemented for NS and ADJ  solution.
Moreover the functional value update in the optimizer has been modified in order to match with the case without the time integration.
The stationary case reading has been moved to the NS import parameters in order to guarantee the possibility to use it also detached the topology optimization procedure.
The default vales of t_end and delta_t have been set to 1e16.
The NS static solution is computed iteratively until the final time is reached.