Added the Stokes solvers  for TopOpt. Corrected the vorticity functional evaluation problem.
The Stokes solver for Physics and Adjoint problem has been implemented by changing a little bit the the one step solver working.
The vorticity functional resulted too big in the functional weights since it wasn't scaled by the viscosity magnitude ( there was an uncommented if statement that now has been commented). In the procedure to fix this bug, also a function to evaluate the vorticity functional starting from the anti-sym part of the gradient has been implemented.
Now that the vorticity functional values have been correctly scaled, it should be possible to use the vorticity term also in multiple constraints case.